fixed small bugs;

+++ sources/application/thread_communication.c
@@ -217,7 +217,7 @@ static int store_ref_gmv_on_flash(void)
         memcpy(&mem_data, p_data, sizeof(uint64_t));
         if (HAL_FLASH_Program(FLASH_TYPEPROGRAM_DOUBLEWORD, addr, mem_data) != HAL_OK)
         {
-            APP_PRINTF("FLASH program error, addr = %d", p_data);
+            APP_PRINTF("FLASH program error, addr = %d", (uint32_t)p_data);
             HAL_FLASH_Lock();
             return -1;
         }
